pyproject.toml
[tool.poetry]
name = "django/es-casestudy"
version = "0.1.0"
description = ""
authors = ["ademir <h.ayberk.demir.34@gmail.com>"]
readme = "README.md"
packages = [{include = "django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.7"
djangorestframework = "^3.15.2"
django-stubs = "^5.1.3"
djangorestframework-stubs = "^3.15.3"
celery = "^5.4.0"
psycopg2-binary = "^2.9.10"
debugpy = "^1.8.13"
elasticsearch = "7.17.0"
docker = "^7.1.0"
drf-yasg = "^1.21.10"
stubs = "^1.0.0"
drf-yasg-stubs = "^0.1.4"
dotenv = "^0.9.9"
celery-stubs = "^0.1.3"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
django-celery-results = "^2.5.1"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.397"
black = "^25.1.0"
flake8 = "^7.1.2"
isort = "^6.0.1"
pylint = "^3.3.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
